+++ changelog.md
@@ -15,6 +15,9 @@ For the latest version of this document, please see [https://github.com/haskell/
 
 * Remove Data.Aeson.Encode module
 
+* `Double` and `Float` infinities are encoded as `"+inf"` and `"-inf"`.
+  Change `To/FromJSONKey` instances to use `"+inf"` and `"-inf"` too.
+
 ### 1.5.6.0
 * Make `Show Value` instance print object keys in lexicographic order.